Fix schema and dev seed.

+++ CommunityFund/db/seeds_dev.rb
@@ -4,7 +4,7 @@ Community.all.each do |c|
   c.update_attributes(:icon => Faker::Avatar.image(c.name, "50x50"))
 end
 
-User.destroy_all
+User.where('username != "administrator"').destroy_all
 
 User.create(
   username: "TimoVink",
